requirements.txt
anyio==4.1.0
attrs==23.1.0
backoff==2.2.1
beartype==0.16.4
cattrs==23.2.3
certifi==2023.11.17
dagger-io==0.9.3
gql==3.4.1
graphql-core==3.2.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
platformdirs==4.1.0
Pygments==2.17.2
rich==13.7.0
sniffio==1.3.0
typing_extensions==4.8.0
yarl==1.9.3
